+++ tests/test_device_subscriptions.py
@@ -127,7 +127,7 @@ class TestDeviceSubscriptions(unittest.TestCase):
             self.assertEqual(devSubscr.notifyToAddress, notifyTo)
             self.assertEqual(devSubscr.notifyRefNodes[0].text, clSubscr.notifyTo_identifier.text)
             self.assertEqual(devSubscr.endToAddress, endTo)
-            self.assertEqual(devSubscr.endToRefNodes[0].text, clSubscr._endTo_identifier.text)
+            self.assertEqual(devSubscr.endToRefNodes[0].text, clSubscr.end_to_identifier.text)
             
             # verify that client subscription object contains the subscription identifier of the device Subscription object
             self.assertEqual(clSubscr.dev_reference_param[0].tag, devSubscr.my_identifier.tag )
@@ -175,7 +175,7 @@ class TestDeviceSubscriptions(unittest.TestCase):
 
             # verify that header contains the identifier of client subscription
             env  = ReceivedSoap12Envelope.fromXMLString(response.as_xml())
-            idents = env.headerNode.findall(namespaces.wseTag('Identifier'))
+            idents = env.headerNode.findall(testSubscr.IDENT_TAG)
             self.assertEqual(len(idents), 1)
             self.assertEqual(idents[0].text, mockstuff.TestDevSubscription.notifyRef)
 
@@ -294,7 +294,7 @@ class TestDeviceSubscriptions(unittest.TestCase):
             self.assertEqual(devSubscr.notifyToAddress, notifyTo)
             self.assertEqual(devSubscr.notifyRefNodes[0].text, clSubscr.notifyTo_identifier.text)
             self.assertEqual(devSubscr.endToAddress, endTo)
-            self.assertEqual(devSubscr.endToRefNodes[0].text, clSubscr._endTo_identifier.text)
+            self.assertEqual(devSubscr.endToRefNodes[0].text, clSubscr.end_to_identifier.text)
 
             # verify that client subscription object contains the subscription identifier of the device Subscription object
             self.assertEqual(clSubscr.dev_reference_param[0].tag, devSubscr.my_identifier.tag)
@@ -361,7 +361,7 @@ class TestDeviceSubscriptions(unittest.TestCase):
 
             # verify that header contains the identifier of client subscription
             env = ReceivedSoap12Envelope.fromXMLString(response.as_xml())
-            idents = env.headerNode.findall(namespaces.wseTag('Identifier'))
+            idents = env.headerNode.findall(testEpisodicSubscr.IDENT_TAG)
             self.assertEqual(len(idents), 1)
             self.assertEqual(idents[0].text, mockstuff.TestDevSubscription.notifyRef)